requirements.txt
gradio==3.36.1
Pillow==10.0.0
torch==2.0.1
ultralytics==8.2.58
numpy==1.24.4
